refactor; added evaluations for new tasks
