+++ data/conll_2012.py
@@ -1,16 +1,7 @@
-import logging
-import re
-import sys
-
 from data.modules import PrepData
+from utils import logger
 
-
-logging.basicConfig(
-    format = '%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-    level = logging.DEBUG,
-    stream = sys.stdout
-)
-logger = logging.getLogger(__name__)
+logger = logger()
 
 def remove_bio_prefixes(label: str) -> str:
     if len(label) < 2 or label[1] != '-':
@@ -52,69 +43,6 @@ class CoNLL2012(PrepData):
                     target = target.rstrip(' ,')
                     target += '), '
                 yield source, target.rstrip(' ,')
-    
-    @staticmethod
-    def unserialize_without_verbs(s: str) -> set[tuple[str, str]]:
-        """
-        Given a sequence, return a set of three tuples of the role-labelled words,
-        with the verb of their respective frame and the role label.
-
-        For example: the string 'eat (A: ham burger, B: chicken), drink (C: coke sprite beer sprite)' will return
-        {
-            (ham, eat, A), (burger, eat, A), (chicken, eat, B),
-            (coke, drink, C), (sprite, drink, C), (beer, drink, C),
-        }
-        """
-        out: set[tuple[str, str]] = set()
-        for verb_frame in re.finditer(r'(\S+) \((.*?\))', s.strip(' ,')):
-            for label_friends in re.finditer(r'([A-Z\-\d]+): (.*?\S)[,)]', verb_frame.group(2)):
-                label, friends = label_friends.group(1), label_friends.group(2)
-                for friend in friends.split():
-                    out.add((friend, label))
-        return out
-        
-    
-    @staticmethod
-    def unserialize(s: str) -> tuple[dict[str, dict[str, dict[str, int]]], int]:
-        """
-        Given a sequence, divide it into verb frames,
-        then for each verb, list the semantic roles and words in the same sentence with that role in a multiset,
-        along with the total number of semantic role labels,
-
-        For example: the string 'eat (A: ham burger, B: chicken), drink (C: coke sprite beer sprite)' will return
-        {
-            'eat_1': {
-                'A': {'ham': 1, 'burger': 1},
-                'B': {'chicken': 1},
-            },
-            'drink_1': {
-                'C': {'coke': 1, 'sprite': 2, 'beer': 1},
-            },
-        },
-        7
-        """
-        out: dict[str, dict[str, dict[str, int]]] = {}
-        verb_counter: dict[str, int] = {}
-        total_labels: int = 0
-        for verb_frame in re.finditer(r'(\S+) \((.*?\))', s.strip(' ,')):
-            verb, bracket_content = verb_frame.group(1), verb_frame.group(2)
-            verb_counter.setdefault(verb, 0)
-            verb_counter[verb] += 1
-            
-            verb = f'{verb}_{verb_counter[verb]}'
-            out.setdefault(verb, {})
-            verb_dict = out[verb]
-            for label_friends in re.finditer(r'([A-Z\-\d]+): (.*?\S)[,)]', bracket_content):
-                label, friends = label_friends.group(1), label_friends.group(2)
-                verb_dict.setdefault(label, {})
-                for friend in friends.split():
-                    friend = friend.strip(' ')
-                    verb_dict[label].setdefault(friend, 0)
-                    verb_dict[label][friend] += 1
-                    total_labels += 1
-            
-        return out, total_labels
-        
 
 if __name__ == '__main__':
     o = CoNLL2012()
